Deliveries Dialog
Adds dialogs for the delivery conversation
