+++ dialogs/__init__.py
@@ -1,3 +1,6 @@
-from .delivery_dialog import DeliveryDialog
+from .cancel_and_help_dialog import CancelAndHelpDialog
+from .create_delivery_dialog import CreateDeliveryDialog
+from .list_deliveries_dialog import ListDeliveriesDialog
+from .main_dialog import MainDialog
 
-__all__ = ["DeliveryDialog"]
+__all__ = ["CancelAndHelpDialog", "CreateDeliveryDialog", "ListDeliveriesDialog", "MainDialog"]
