+++ dialogs/constants/__init__.py
@@ -1,0 +1,4 @@
+from .actions import Actions
+from .dialog import Dialog
+
+__all__ = ["Actions", "Dialog"]
