+++ dialogs/constants/actions.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class Actions(Enum):
+    CREATE_DELIVERY: str = "Create delivery"
+    EXIT: str = "Exit"
+    LIST_DELIVERIES: str = "List deliveries"
