+++ recognizers/__init__.py
@@ -1,3 +1,3 @@
-from .movie_recognizer import MovieRecognizer
 
-__all__ = ["MovieRecognizer"]
+
+__all__ = [""]
